requirements.txt
autopep8==1.6.0
coreapi==2.3.3
Django==3.2.9
PyYAML==6.0
django-extensions==3.1.5
djangorestframework==3.12.4
drf-spectacular==0.21.0
openpyxl==3.0.9
pandas==1.3.4
pyinstaller==4.6
pylint-django==2.4.4
pylint==2.11.1
uritemplate==4.1.1
waitress==2.0.0
whichcraft==0.6.1
whitenoise==5.3.0
xlrd==2.0.1
fpdf==1.7.2
Eel==0.14.0
appdirs==1.4.4
django-cryptography==1.1
